fix: Make DOCKER_CONFIG available to buildctl
When povisioning a finch VM, the CLI inserts a line into the
user's bash rc that points DOCKER_CONFIG to the user's finch dir.
Finch-daemon running through systemd doesn't load the bash rc
and so it doesn't get this config. As a result, build requests
that go through finch-daemon don't pass this variable to buildctl
so buildctl/buildkit is unable to build on base images in private
repositories.

This change temporarily fixes this by inserting DOCKER_CONFIG into
the finch-daemon process based on the owner of the socket. That
way, buildctl is able to load the finch config and retrieve credentials.

This should eventually be replaced by somehow passing the credentials
that come in to the build API directly to buildctl rather than loading
credentials based on the config.

Signed-off-by: Kern Walster <walster@amazon.com>