ci: add samcli tests through lima vm on codebuild (#331)
This is an implementation of automated tests for SAMcli against Finch. It runs in a CodeBuild Mac runner so that it may start the Lima VM. Due to issues in the CodeBuild environment this test suite must run in sequence, taking about 1 hour 45 minutes to complete. This will therefore be run nightly, with the option for manual trigger if needed.

---------

Signed-off-by: ayush-panta <ayushkp@amazon.com>