ci: better teardown failure detection in samcli-vm (#344)
An issue with flaky tests due to teardown errors appeared after a recent merge. I have added a simple approach that checks test logs for a few patterns that indicate a teardown failure. If there are unexpected tests, we will only prompt the test to fail if there were no teardown issues.

Signed-off-by: ayush-panta <ayushkp@amazon.com>